Rename typography folder to text
