+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/index.ts
@@ -13,4 +13,4 @@ export * from './main-header';
 export * from './navigation-header';
 export * from './spinner';
 export * from './theme';
-export * from './typography';
+export * from './text';
