+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/text/FootnoteMini.tsx
